+++ CHANGELOG.md
@@ -7,6 +7,8 @@ Inspired from [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 ### Features
 ### Enhancements
 ### Bug Fixes
+- Avoid race condition putting the same document id in DDB Client ([#228](https://github.com/opensearch-project/opensearch-remote-metadata-sdk/pull/228))
+
 ### Infrastructure
 ### Documentation
 ### Maintenance